Add postgres build flag

+++ migrations/20230521061658_price_value_attestations.up.sql
@@ -1,8 +1,8 @@
 -- Add migration script here
 CREATE TABLE price_value_attestations (
     validator_public_key text NOT NULL,
-    slot_number integer NOT NULL,
-    value integer NOT NULL,
+    slot_number BIGINT NOT NULL,
+    value BIGINT NOT NULL,
     signature text NOT NULL,
     PRIMARY KEY (validator_public_key, slot_number)
 );
