+++ migrations/20230521105847_price_interval_attestations.up.sql
@@ -2,7 +2,7 @@
 CREATE TABLE price_interval_attestations (
     validator_public_key text NOT NULL,
     signature text NOT NULL PRIMARY KEY,
-    slot_number integer NOT NULL,
-    value integer NOT NULL,
-    interval_size integer NOT NULL
+    slot_number  BIGINT NOT NULL,
+    value  BIGINT NOT NULL,
+    interval_size  BIGINT NOT NULL
 );
