+++ migrations/20230522052218_aggregated_interval_attestations.up.sql
@@ -2,8 +2,8 @@
 CREATE TABLE aggregate_interval_attestations (
     aggregate_signature text NOT NULL PRIMARY KEY,
     aggregate_public_key text NOT NULl,
-    slot_number integer NOT NULL,
-    value integer NOT NULL,
-    interval_size integer NOT NULL,
-    num_validators integer NOT NULL
+    slot_number  BIGINT NOT NULL,
+    value  BIGINT NOT NULL,
+    interval_size  BIGINT NOT NULL,
+    num_validators  BIGINT NOT NULL
 );
